First try @ configurable menu extension - refs #451
Herein is implemented an attempt at allowing users to add
menu items to the main menu in Mirador via a configuration
setting.  Plaintext links to href or that trigger callbacks
are more or less fully handled.  Sublists and explicit handling
of links with icons are the next step here.

index.html contains a basic example of each type of link, and
there is inline documentation co-located with the functions that
implement it.